extension/backend/requirements.txt
aiohttp
fastapi
idna
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith
motor
multidict
pydantic
PyJWT
pymongo
pypdf
python-dotenv
python-multipart
urllib3
uvicorn
virtualenv
nest-asyncio
beautifulsoup4
mangum
faiss-cpu